VERSION 1.0 ESTABLE Cambio en BD a nivel de tipo de dato (Telefono) y script general. Eliminación de archios NO VITALES para ser compartidos

+++ src/java/BusinessLogic/Controller/HandleUser.java
@@ -15,7 +15,7 @@ import javax.faces.context.FacesContext;
  */
 public class HandleUser {
 
-    public String doCreate(String name, String lastname, int age, String address, double phone, String email, String username, String password1, String password2, String role) {
+    public String doCreate(String name, String lastname, int age, String address, long phone, String email, String username, String password1, String password2, String role) {
         User user = new User();
         Role roleObject = new Role(Integer.parseInt(role));
         user.setName(name);
